Rename modeling.py to modelling.py & Update example_config.yaml
